Ajout de deux attaques avec puissance différentes. Gestion d'une puissance différente selon le type d'attaque et le type de pokemon. Ajout de la classe Alternance, à gérer

+++ Pokemon_Stadium/src/main/scala/pokemon_stadium/Defense.scala
@@ -1,6 +1,6 @@
 package pokemon_stadium
 
-class Defense(nom: String, defense: Defense, precision: Float, puissance: Float) {
+class Defense(nom: String, defense: Defense, precision: Double, puissance: Double) {
 
 }
